Add button on container info pane to show only that container.
Adds a button, located on the info pane for each container, to show only that container.
This button will hide other containers, if they have visible tabs, and show the container the button was pressed for, if it has hidden tabs.

This is a much-needed function to effectively use containers as workspaces, and will please most people coming to multi-account-containers from Panorama tabs.