Open new tabs in same container as previous tab.
This commit adds functionality that will try to open new tabs in the same container as the previously active tab in that window, if the new tab doesn't already have a container, and the previously active tab did.
Some undocumented / implicity functionality in the webextension API is used:
 * `onCreated` is called for a new tab before `onActivated` for that tab is called
 * `onUpdated` will be called for all new tabs, also `about:newtab` and the like (sets the favicon)
 * The first `onUpdated` for a new tab, that is not an `about:`-tab, will set the URL

It is a little bit wonky, as the original new tab is created and displayed (although not loaded) before it can be closed a new tab can be created. However, it is the best workaround I can find to add this functionality until Bugzilla 1406371 is solved.

This would fix #462, #448 and #406 (I think)
Also relates to #943 and #544