list sites which always load with container
Thanks for building this addon!  It's really great and I really love the
functionality it provides.  I have a small change that I'd like to make.
It's motivated by my personal frustration with trying to figure out what
is supposed to load in a specific container.

The data storage system used is optimized for quick lookups on webpage
load, which is the right choice.  Instead of changing that, I decided
instead to iterate through the full data storage and pick out the
relevant items and trace them back to the container that they belong to
for display.  I suspect that with the size of the data storage, it's
probably fine.  Other than building a mapping in a different key of the
local storage when adding a site to a container, I don't see any way to
make this quicker on lookup.

I'm not at all an extension developer, nor do I write a lot of html or
css.  I probably butchered the display portion of this, but it mostly
works...  I would not expect this patch to merge in this state, but I'd
be interested in working on making it possible to merge.  Maybe hiding
this functionality behind an addon preference?

If this functionality isn't desirable in the official addon, is there
anything which would block me from submitting a fork to
addons.mozilla.org?  It's primarily because I'd like to use it on a
release copy of firefox, and so I'd need it signed by addons.