Add allowlist for isolated containers
This implements a per-container allow-list that is applied to isolated
containers. This can allow multiple containers to have an overlapping set of
sites that they are isolated to.

For example: Container A and Container B could both be isolated and have
calendar.example.com in their allow list. Both Container A and Container B
would then be able to visit calendar.example.com.

See: https://github.com/mozilla/multi-account-containers/issues/1892 and
https://github.com/mozilla/multi-account-containers/issues/1887 for additional
motivation.